Copy right changes

+++ ximple-dev/com/ximpleware/xpath/NumExpr.java
@@ -1,5 +1,5 @@
 /* 
- * Copyright (C) 2002-2004 XimpleWare, info@ximpleware.com
+ * Copyright (C) 2002-2006 XimpleWare, info@ximpleware.com
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -27,8 +27,8 @@ public class NumExpr extends Expr{
 		dval = d;
 	}
 	public String toString(){
-		if (dval == (int) dval){
-			return ""+(int) dval;
+		if (dval == (long) dval){
+			return ""+(long) dval;
 		}
 		return  ""+dval;
 	}
